pyproject.toml
[project]
name = "rag-chatbot"
version = "1.0.0"
description = "RAG Chatbot using AWS Bedrock, ChromaDB, FastAPI, LangChain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "gunicorn>=22.0.0",

    # LangChain
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",

    # AWS
    "boto3>=1.42.78",

    # Vector DB
    "chromadb>=0.5.0",

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "cryptography>=42.0.0",

    # Config
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",

    # Document loaders
    "pypdf>=4.0.0",
    "unstructured>=0.14.0",

    # Cache
    "cachetools>=5.3.0",

    # OpenTelemetry
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",

    # File upload
    "python-multipart>=0.0.9",
]

[tool.hatch.build.targets.wheel]
packages = ["Src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
